Use BUILD_CC, HOST_CC, TARGET_CC, and TARGET_LD
TARGET_CC (replacing the CROSS_COMPILE variable) is always a m68k
compiler, if set at all, to compile code for the Atari ST. It's
needed to build PSGPLAY.TOS, or to rebuild lib/tos/tos (the boot
ROM file) which for convenience has been precompiled in the repo.
TARGET_LD is the linker needed for linker scripts.

For most people, TARGET_CC is therefore optional. The target is
the m68k architecture fundamentally being emulated, in every case.

The host, corresponding to HOST_CC, is the architecture performing
(running) the emulation, which often but not always is the same
as the build architecture, corresponding to BUILD_CC.